==> 2. data preprocessing-2.ipynb <==
pip install miceforest

import pandas as pd
import numpy as np
# import miceforest as mf
from scipy import stats
from imblearn.over_sampling import SMOTE
from collections import Counter

mf.__version__

data = pd.read_csv("./output/datalong_en_024.csv")

state_feature_list = [
    'gender',
    'age',
    'age_at_diagnosis',
    'weight',
    'height',
    'ABS',
    'BMI',

    'previous_radiation',
    'previous_chemotherapy',

    'congestive_heart_failure',
    'angina_pectoris',
    'myocardial_infarction',
    'copd',
    'peptic_ulcer',
    'thyroid_disease',
    'diabetes',
    'autoimmune_disease',
    'acquired_immune_deficiency_syndrome',

    'tumor_stage_pM',
    'tumor_stage_pN',
    'tumor_stage_pT',

    'PNI'
]

dynamic_feature_list = [
    'GFR',
    'AST',
    'TBIL',
    'PA',
    'ALP',
    'ALB',
    'ALT',
    'Crea',

    'PLT',
    'Hb',
    'WBC',
    'Lymph',
    'Neut',
    
    'prevention',

    'PS_score'
]

# data imputation

## chronic disease 
慢性病比例较低 处理为diabetes(糖尿病) /peptic_ulcer（消化性溃疡）/ other_chronic（除上述外的其他慢性病）

data["other_chronic"] = data[["congestive_heart_failure", "angina_pectoris", "myocardial_infarction", "copd", "thyroid_disease", 
                              "autoimmune_disease", "acquired_immune_deficiency_syndrome"]].sum(axis = 1)
data["other_chronic"] = data["other_chronic"].map(lambda x: 1 if x >= 1 else 0)

## TNM tumor stage

pM stage：all M0, drop this column 
pN stage：pNx:imputated with pN0; nan: imputed by miceforest  
pT stage：nan: imputed by miceforest   

data["tumor_stage_pN"] = data["tumor_stage_pN"].map(lambda x: "pN0" if x == "pNx" else x)

## ps score
只保留基线

data.rename(columns = {"C1*PS_score": "PS_score"}, inplace = True)

## output variables with missing values

state_feature_list = [
    'gender',
    'age_at_diagnosis',
    'weight',
    'previous_radiation',
    'previous_chemotherapy',
    'peptic_ulcer',
    'diabetes',
    'other_chronic',
    'tumor_stage_pN',
    'tumor_stage_pT',
    'PNI',
    "PS_score"
]

temp_missing = data[["patient_sn", "visit_year", "event", "event#cycle", "cycle_num"] + \
                    state_feature_list +\
                    [name for name in data.columns.to_list() if "C" in name and not "PS_score" in name]]

将数据集整合成填补后一致的结构，不符合患者周期对应的特征取值为np.nan

data_missing = temp_missing.copy()
cycle_list = ['GFR', 'AST', 'TBIL', 'PA', 'ALP', 'ALB', 'ALT', 'Crea', 'PLT', 'Hb', 'WBC', 'Lymph', 'Neut', 'prevention', 'interval']
for i in range(len(data_missing)):
    event_num = data_missing["event#cycle"][i]
    cycle_num = data_missing["cycle_num"][i]
    if np.isnan(event_num):
        num = cycle_num
    else:
        num  = min(event_num, cycle_num)
    num = int(num)
    for j in range(num+1, 9):
        pre = "C" + str(j) + "*"
        for name in cycle_list:
            data_missing[pre+name][i] = np.nan
data_missing

# data_missing.to_csv("./output/missing.csv", index  = False)

# multiple imputation

静态指标mice一个，得到填补后的静态数据集后再逐周期填补。

## static variables

# data_missing = pd.read_csv("./output/missing.csv")

state_missing = data_missing[state_feature_list]
state_missing.isnull().sum()

for col in state_missing.columns:
    print(col + " : " + str(state_missing[col].dtype))

for col_name in ['tumor_stage_pN', 'tumor_stage_pT']:
    state_missing[col_name] = state_missing[col_name].astype("category")


kernel0 = mf.ImputationKernel(
    state_missing,
    datasets = 1, # number of final completed datasets
    save_all_iterations = True,
    save_models = 1, # whether the model calculating the missing values is saved =0:unsaved; =1:save the final model; =2:save all models
    random_state = 0
)

kernel0.mice(
    iterations = 10,
    # n_jobs = -1
)

print(kernel0)

state_complete = kernel0.complete_data(0) 
state_complete

# replace the original variables with imputated variables
data_missing_t = data_missing.copy()
for feature in list(state_complete.columns):
    # print(feature)
    data_missing_t[feature] = list(state_complete[feature])
data_missing_t

data_complete0 = data_missing_t[["patient_sn", "visit_year", "event", "event#cycle", "cycle_num"]+ \
                                state_feature_list+ \
                                [name for name in data_missing_t.columns if "prevention" in name]
                               ]

## dynamic features

state_feature_list = [
    'gender',
    'age_at_diagnosis',
    'weight',
    'previous_radiation',
    'previous_chemotherapy',
    'peptic_ulcer',
    'diabetes',
    'other_chronic',
    'tumor_stage_pN',
    'tumor_stage_pT',
    'PNI',
    'PS_score'
]
dynamic_feature_list = ["GFR", "AST",  "TBIL", "PA", "ALP", "ALB", "ALT", "Crea", "PLT", "Hb", "WBC", "Lymph", "Neut"]

def dynamic_mice(num):
    cycle = "C"+str(num)
    cycle_t = data_missing_t[(data_missing_t["cycle_num"] >= num) & \
                             ((data_missing_t["event#cycle"] >= num) | (np.isnan(data_missing_t["event#cycle"])))]
    print(len(cycle_t))
    cycle_missing = cycle_t[state_feature_list + \
                           [cycle+"*"+name for name in dynamic_feature_list]]
    
    for col in cycle_missing.columns:
        if cycle_missing[col].dtype == np.object:
            cycle_missing[col] = cycle_missing[col].astype("category")
    
    kernel = mf.ImputationKernel(
        cycle_missing,
        datasets = 1, # number of final completed datasets
        save_all_iterations = True,
        save_models = 1, # whether the model calculating the missing values is saved =0:unsaved; =1:save the final model; =2:save all models
        random_state = 0
    )
    
    kernel.mice(
        iterations = 15,
        n_jobs = -1
    )
    
    data_complete = kernel.complete_data(0) # which completed datasets is chosen
    data_complete = data_complete[[cycle+"*"+name for name in dynamic_feature_list]]
    
    data_complete["patient_sn"] = list(cycle_t["patient_sn"])
    data_complete[cycle+"*interval"] = list(cycle_t[cycle+"*interval"])
    
    return kernel, data_complete

# data_complete0
kernel1,  data_complete1 = dynamic_mice(1)
data_complete01 = pd.merge(data_complete0, data_complete1, on = "patient_sn", how = "left")

kernel2,  data_complete2 = dynamic_mice(2)
data_complete012 = pd.merge(data_complete01, data_complete2, on = "patient_sn", how = "left")

kernel3,  data_complete3 = dynamic_mice(3)
data_complete0123 = pd.merge(data_complete012, data_complete3, on = "patient_sn", how = "left")

kernel4,  data_complete4 = dynamic_mice(4)
data_complete01234 = pd.merge(data_complete0123, data_complete4, on = "patient_sn", how = "left")

kernel5,  data_complete5 = dynamic_mice(5)
data_complete012345 = pd.merge(data_complete01234, data_complete5, on = "patient_sn", how = "left")

kernel6,  data_complete6 = dynamic_mice(6)
data_complete0123456 = pd.merge(data_complete012345, data_complete6, on = "patient_sn", how = "left")

kernel7,  data_complete7 = dynamic_mice(7)
data_complete01234567 = pd.merge(data_complete0123456, data_complete7, on = "patient_sn", how = "left")

kernel8,  data_complete8 = dynamic_mice(8)
data_complete012345678 = pd.merge(data_complete01234567, data_complete8, on = "patient_sn", how = "left")

data_complete012345678

data_complete012345678.to_csv("./output/complete.csv", index = False)

## after imputation

basic_list = ["GFR", "AST",  "TBIL", "PA", "ALP", "ALB", "ALT", "Crea", "PLT", "Hb", "WBC", "Lymph", "Neut"]
cycle_list = list()
for i in range(1, 9):
    cycle_list = cycle_list + ["C"+str(i)+"*"+name for name in basic_list]

result_list = list()
for column_name in cycle_list:
    x1 = list(data_missing[~data_missing[column_name].isnull()][column_name])
    x2 = list(data_complete012345678[~data_complete012345678[column_name].isnull()][column_name])
    result_list.append(
        {
            "feature": column_name,
            "pvalue": stats.stats.ttest_ind(x1, x2).pvalue
        }
    )

pd.DataFrame(result_list).sort_values(["pvalue"], ascending=True)

# final table (long type)

import pandas as pd
import numpy as np
import miceforest as mf
from scipy import stats

data_complete = pd.read_csv("./output/填补之后的_complete.csv")
data = pd.read_csv("./output/datalong_en_024.csv")
data_prevention = data[["patient_sn"]+[name for name in data.columns if "prevention" in name]]
data_complete = pd.merge(data_complete, data_prevention, on = "patient_sn")

data_complete["id"] = data_complete.index
state_feature = [
    'patient_sn',
    'id',
    'cycle_num',
    'visit_year',
    'gender',
    'age_at_diagnosis',
    'weight',
    'previous_radiation',
    'previous_chemotherapy',
    'peptic_ulcer',
    'diabetes',
    'other_chronic',
    'tumor_stage_pN',
    'tumor_stage_pT',
    'PNI',
    'PS_score'
]

data_state = data_complete[state_feature]

data_complete[["event#cycle","cycle_num"]]

dynamic_feature_list = ["GFR", "AST",  "TBIL", "PA", "ALP", "ALB", "ALT", "Crea", "PLT", "Hb", "WBC", "Lymph", "Neut", "prevention", "interval"]

dynamic_result = list()
for i in range(len(data_complete)):
    event_cycle = data_complete["event#cycle"][i]
    num_cycle = data_complete["cycle_num"][i]
    if np.isnan(event_cycle):
        cnt = num_cycle
    else:
        cnt = min(event_cycle, num_cycle)
    cnt = int(cnt)
    
    for num in range(1, cnt+1):
        t_dict = dict()
        
        cycle_num = "C"+str(num)+"*"
        t_dict["patient_sn"] = data_complete["patient_sn"][i]
        t_dict["cycle"] = num
        
        for cycle_feature in dynamic_feature_list:
            if cycle_feature == "interval":
                t_dict[cycle_feature] = data_complete["C"+str(num)+"*"+cycle_feature][i]
            elif cycle_feature == "prevention" and num == 1:
                t_dict[cycle_feature] = 0
            else:
                t_dict[cycle_feature] = data_complete["C"+str(num-1)+"*"+cycle_feature][i]
        
        # event    
        if np.isnan(event_cycle):
            t_dict["event"] = 0
        elif num == event_cycle:
            t_dict["event"] = 1
        else:
            t_dict["event"] = 0
            
        # previous cycle interval
        if num == 1:
            t_dict["last_interval"] = 30
        else:
            t_dict["last_interval"] = data_complete["C"+str(num-1)+"*"+"interval"][i]
        
        dynamic_result.append(t_dict)
data_dynamic = pd.DataFrame(dynamic_result)

# the start-time and end-time of present cycle of chemotherapy
data_dynamic["time2"] = data_dynamic.groupby(["patient_sn"])["interval"].cumsum()
data_dynamic["time1"] = data_dynamic["time2"] - data_dynamic["interval"]
data_dynamic.drop(["interval"], axis = 1, inplace = True)
for i in range(len(data_dynamic)):
    if data_dynamic["time1"][i] == data_dynamic["time2"][i]:
        data_dynamic["time2"][i] = data_dynamic["time1"][i]+1

data_long = pd.merge(data_dynamic, data_state, on = "patient_sn", how = "left")
# data_long.to_csv("./output/data_long.csv", index = False)
data_long

# data discretization

data_long = pd.read_csv("./output/data_long.csv")
# most of PS_score is 0, merge values greater than 0 into one categroy
data_long["PS_score_gt_0"] = data_long["PS_score"].map(lambda x: 1 if x == 1 else 0)

data_long["tumor_stage_pN"] = data_long["tumor_stage_pN"].map(lambda x: 0 if x == "pN0" else 1)

continous_feature = [
    "age_at_diagnosis",
    "weight",
    "PNI",
    "GFR",  "TBIL", "ALP", "ALB", "ALT", "Crea", "PLT", "Hb", "WBC",
    "AST", "PA", "Lymph", "Neut"
]

def quartile_feature(feature_name, data):
    """
    将连续变量根据四分位数生成取值1-4的连续变量
    """
    df = data.copy()
    if feature_name == "age_at_diagnosis":
        df[feature_name] = pd.cut(df[feature_name], [0, 60, 120], labels = ["le_60", "lt_60"])
        df["age_at_diagnosis_gt_60"] = df["age_at_diagnosis"].map(lambda x: 1 if x == "lt_60" else 0)
    else:
        quantile_list = df[feature_name].quantile([0, 0.25, 0.5, 0.75, 1]).to_list()
        quantile_list[0] = 0.99*quantile_list[0]
        df["{}_q".format(feature_name)] = pd.cut(df[feature_name], quantile_list, labels = ["1", "2", "3", "4"], right = True)
        df["{}_q".format(feature_name)] = df["{}_q".format(feature_name)].map(lambda x: int(x))
    return df

df = data_long
for feature in continous_feature:
    df = quartile_feature(feature, df)

data_long.to_csv("./output/data_long_quartile_feature.csv", index = False)

df_new = df[[
    'patient_sn',
    'visit_year',
    'cycle_num',
    'cycle',
    'event',

    'gender',
    'tumor_stage_pN',

    'last_interval',
    'prevention',  
    'GFR_q',
    'TBIL_q',
    'ALP_q',
    'ALB_q',
    'ALT_q',
    'Crea_q',
    'PLT_q',
    'Hb_q',
    'WBC_q',
    'AST_q',
    'PA_q',
    'Lymph_q',
    'Neut_q',df_new.to_csv("./output/data_valid.csv", index = False)
    
]]


# Smote

data = pd.read_csv("/Users/weiyuna/Desktop/工作/合作项目/医大一-胃癌/模型/code_202203/output/data_valid.csv")

data_valid = data
i = 8
data_valid = data_valid[data_valid.cycle <= i]

data_train = data_valid[data_valid["visit_year"] <= 2018]
X_train = data_train[[x for x in data.columns if x not in ["patient_sn", "visit_year", "cycle_num", "event"]]]
y_train = data_train["event"]

data_test = data_valid[data_valid["visit_year"] > 2018]
data_test.reset_index(inplace = True)
X_test = data_test[[x for x in data.columns if x not in ["patient_sn", "visit_year", "cycle_num", "event"]]]
y_test = data_test["event"]

smo = SMOTE(random_state = 0)
X_train_smo, y_train_smo = smo.fit_resample(X_train, y_train)
smote_train = X_train_smo
smote_train["event"] = y_train_smo
smote_train["target"] = "train"
print('smote train: {}'.format(Counter(y_train_smo)))

X_test_smo, y_test_smo = smo.fit_resample(X_test, y_test)
smote_test = X_test_smo
smote_test["event"] = y_test_smo
smote_test["target"] = "test"
print('smote test: {}'.format(Counter(y_test_smo)))

smote_data = smote_train.append(smote_test)
# smote_data.to_csv("smote_data.csv", index = False)
